+++ js/default.js
@@ -1,3 +1,55 @@
+Vue.component("roxyfileman",{
+    template:"<bs-input/>",
+    mounted(){
+        this.$nextTick(function(){
+            $(this.$el).on("click", function () {
+                var d = new Date().getTime();
+                d = "roxy_" + d;
+
+                $(this).attr("data-roxy-id", d);
+
+                var resize_width = $(this).attr('roxy-resize-width');
+                var type = ($(this).attr('roxy-type') == undefined) ? "image" : $(this).attr('roxy-type');
+
+                var session = "";
+                if ($(this).attr("roxy-session")) {
+                    session = $(this).attr("roxy-session");
+                }
+
+                var path = "plugins/RoxyFileman.1.4.5/fileman/index.html?type=" + type + "&integration=custom&txtFieldId=" + d + "&session=" + session;
+
+                if (resize_width != undefined) {
+                    path += "&resize_width=" + resize_width;
+                }
+                $.fancybox.open({
+                    src: path,
+                    type: "iframe"
+                });
+            });
+        });
+
+    }
+});
+
+Vue.component("ckeditor",{
+    template:"<textarea><slot></slot></textarea>",
+    mounted(){
+        this.$nextTick(function(){
+            if (typeof CKEDITOR != 'undefined') {
+                
+                
+                var base = $("base").attr("href");
+                CKEDITOR.config.filebrowserImageBrowseUrl = base + "plugins/RoxyFileman.1.4.5/fileman/index.html?type=image";
+                CKEDITOR.config.filebrowserBrowseUrl = base + "plugins/RoxyFileman.1.4.5/fileman/index.html";
+    
+                CKEDITOR.replace(this.$el);
+            }
+        });
+
+    }
+});
+
+
 var vue_init = function () {
     document.querySelectorAll('alt-box').forEach(o => {
         new Vue({
@@ -320,10 +372,11 @@ document.addEventListener("DOMContentLoaded", function(event) {
             CKEDITOR.config.filebrowserImageBrowseUrl = base + "plugins/RoxyFileman.1.4.5/fileman/index.html?type=image";
             CKEDITOR.config.filebrowserBrowseUrl = base + "plugins/RoxyFileman.1.4.5/fileman/index.html";
 
-            $("textarea.ckeditor").not("._ckeditor").each(function (i, o) {
+            
+            /*$("textarea.ccc").not("._ckeditor").each(function (i, o) {
                 $(o).addClass("_ckeditor");
                 CKEDITOR.replace(o);
-            });
+            });*/
         }
         //--------------